remove xorTwoIds
